Streamline CIV creation
