+++ app/grandchallenge/reader_studies/tasks.py
@@ -8,13 +8,11 @@ from grandchallenge.components.models import (
     ComponentInterface,
     ComponentInterfaceValue,
 )
-from grandchallenge.notifications.models import Notification, NotificationType
 from grandchallenge.reader_studies.models import (
     Answer,
     DisplaySet,
     ReaderStudy,
 )
-from grandchallenge.uploads.models import UserUpload
 
 
 @transaction.atomic
@@ -49,92 +47,6 @@ def add_scores_for_display_set(*, instance_pk, ds_pk):
             display_set=display_set,
         ).get()
         add_score(instance, ground_truth.answer)
-
-
-@shared_task(
-    **settings.CELERY_TASK_DECORATOR_KWARGS["acks-late-micro-short"],
-)
-def add_image_to_display_set(
-    *, display_set_pk, interface_pk, upload_session_pk=None, image_pk=None
-):
-    with transaction.atomic():
-        display_set = DisplaySet.objects.get(pk=display_set_pk)
-
-        if upload_session_pk:
-            upload_session = RawImageUploadSession.objects.get(
-                pk=upload_session_pk
-            )
-            try:
-                image = Image.objects.get(origin_id=upload_session_pk)
-            except (Image.DoesNotExist, Image.MultipleObjectsReturned):
-                error_message = "Image imports should result in a single image"
-                upload_session.status = RawImageUploadSession.FAILURE
-                upload_session.error_message = error_message
-                upload_session.save()
-                return
-
-        if image_pk:
-            image = Image.objects.filter(pk=image_pk).get()
-
-        interface = ComponentInterface.objects.get(pk=interface_pk)
-
-        display_set.values.remove(
-            *display_set.values.filter(interface=interface)
-        )
-        civ, created = ComponentInterfaceValue.objects.get_or_create(
-            interface=interface, image=image
-        )
-        if created:
-            try:
-                civ.full_clean()
-            except ValidationError as e:
-                # this should only happen for new uploads
-                upload_session.status = RawImageUploadSession.FAILURE
-                upload_session.error_message = e.message
-                upload_session.save()
-        display_set.values.add(civ)
-
-
-@shared_task(**settings.CELERY_TASK_DECORATOR_KWARGS["acks-late-2xlarge"])
-def add_file_to_display_set(
-    *,
-    user_upload_pk,
-    display_set_pk,
-    interface_pk,
-    civ_pk=None,
-):
-    user_upload = UserUpload.objects.get(pk=user_upload_pk)
-    display_set = DisplaySet.objects.get(pk=display_set_pk)
-    interface = ComponentInterface.objects.get(pk=interface_pk)
-    error = None
-    with transaction.atomic():
-        civ = ComponentInterfaceValue.objects.create(interface=interface)
-        try:
-            civ.validate_user_upload(user_upload)
-            civ.full_clean()
-        except ValidationError as e:
-            transaction.set_rollback(True)
-            error = str(e)
-        else:
-            user_upload.copy_object(to_field=civ.file)
-            civ.save()
-            display_set.values.add(civ)
-            if civ_pk is not None:
-                # Remove the preiously assigned civ from the display set
-                civ = ComponentInterfaceValue.objects.get(pk=civ_pk)
-                display_set.values.remove(civ)
-
-    if error is not None:
-        Notification.send(
-            kind=NotificationType.NotificationTypeChoices.FILE_COPY_STATUS,
-            actor=user_upload.creator,
-            message=f"File for interface {interface.title} failed validation.",
-            target=display_set.reader_study,
-            description=(
-                f"File for interface {interface.title} added to {display_set_pk} "
-                f"in {display_set.reader_study.title} failed validation:\n{error}."
-            ),
-        )
 
 
 @shared_task(**settings.CELERY_TASK_DECORATOR_KWARGS["acks-late-2xlarge"])
