+++ app/tests/components_tests/test_tasks.py
@@ -4,8 +4,11 @@ import pytest
 from celery.exceptions import MaxRetriesExceededError
 
 from grandchallenge.algorithms.models import AlgorithmImage
+from grandchallenge.cases.models import RawImageUploadSession
+from grandchallenge.components.models import ComponentInterfaceValue
 from grandchallenge.components.tasks import (
     _retry,
+    add_image_to_object,
     civ_value_to_file,
     encode_b64j,
     execute_job,
@@ -17,9 +20,14 @@ from tests.algorithms_tests.factories import (
     AlgorithmFactory,
     AlgorithmImageFactory,
 )
-from tests.components_tests.factories import ComponentInterfaceValueFactory
+from tests.cases_tests.factories import RawImageUploadSessionFactory
+from tests.components_tests.factories import (
+    ComponentInterfaceFactory,
+    ComponentInterfaceValueFactory,
+)
 from tests.evaluation_tests.factories import MethodFactory
-from tests.factories import WorkstationImageFactory
+from tests.factories import ImageFactory, WorkstationImageFactory
+from tests.reader_studies_tests.factories import DisplaySetFactory
 
 
 @pytest.mark.django_db
@@ -222,3 +230,53 @@ def test_upload_to_registry_and_sagemaker(
     image = AlgorithmImage.objects.get(pk=image.pk)
     assert image.is_in_registry
     assert image.is_desired_version
+
+
+@pytest.mark.django_db
+def test_add_image_to_object(settings):
+    settings.task_eager_propagates = (True,)
+    settings.task_always_eager = (True,)
+
+    ds = DisplaySetFactory()
+    us = RawImageUploadSessionFactory()
+    ci = ComponentInterfaceFactory(kind="IMG")
+
+    error_message = "Image imports should result in a single image"
+
+    add_image_to_object(
+        app_label=ds._meta.app_label,
+        model_name=ds._meta.model_name,
+        upload_session_pk=us.pk,
+        object_pk=ds.pk,
+        interface_pk=ci.pk,
+    )
+
+    assert ComponentInterfaceValue.objects.filter(interface=ci).count() == 0
+    us.refresh_from_db()
+    assert us.status == RawImageUploadSession.FAILURE
+    assert us.error_message == error_message
+
+    im1, im2 = ImageFactory.create_batch(2, origin=us)
+
+    add_image_to_object(
+        app_label=ds._meta.app_label,
+        model_name=ds._meta.model_name,
+        upload_session_pk=us.pk,
+        object_pk=ds.pk,
+        interface_pk=ci.pk,
+    )
+    assert ComponentInterfaceValue.objects.filter(interface=ci).count() == 0
+    us.refresh_from_db()
+    assert us.status == RawImageUploadSession.FAILURE
+    assert us.error_message == error_message
+
+    im2.delete()
+
+    add_image_to_object(
+        app_label=ds._meta.app_label,
+        model_name=ds._meta.model_name,
+        upload_session_pk=us.pk,
+        object_pk=ds.pk,
+        interface_pk=ci.pk,
+    )
+    assert ComponentInterfaceValue.objects.filter(interface=ci).count() == 1
